eliminate crawler.js
